\name{rkhscovfun-package}
\alias{rkhscovfun-package}
\alias{rkhscovfun}
\docType{package}
\title{
Nonparametric Operator-Regularized Covariance Function Estimation
}
\description{
This package provides a class of nonparametric covariance function estimators that utilize spectral regularization of an operator.
}
\details{
\tabular{ll}{
Package: \tab rkhscovfun\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2018-02-16\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
  Raymond K. W. Wong and Xiaoke Zhang
}
\references{R. K. W. Wong and X. Zhang. (2017) "Nonparametric Operator-Regularized Covariance Function Estimation for Functional Data".}
\examples{
}
